backend/requirements.txt
absl-py
aiofiles
annotated-types
anyio
astunparse
audioread
bcrypt
blinker
cachetools
certifi
cffi
charset-normalizer
click
colorama
decorator
dnspython
email_validator
fastapi
ffmpeg
Flask
flatbuffers
gast
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
google-pasta
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
h5py
httplib2
idna
itsdangerous
Jinja2
joblib
keras
lazy_loader
libclang
librosa
llvmlite
Markdown
markdown-it-py
MarkupSafe
mdurl
ml_dtypes
motor
msgpack
namex
numba
numpy
opt_einsum
optree
packaging
passlib
platformdirs
pooch
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pydub
Pygments
PyJWT
pymongo
pyparsing
python-multipart
requests
rich
rsa
scikit-learn
scipy
setuptools
six
sniffio
soundfile
soxr
SQLAlchemy
starlette
tensorboard
tensorboard-data-server
tensorflow
termcolor
threadpoolctl
tqdm
typing_extensions
uritemplate
urllib3
uvicorn
Werkzeug
wheel
wrapt
